Editor de niveles, coloca diferentes assets, no controla la superposición de assets ni que queden fuera de la cuadrícula

+++ ProyectoFinalCurso/Assets/LevelEditor/Scripts/GridGenrator.cs
@@ -6,7 +6,7 @@ public class GridGenrator : MonoBehaviour
 {
     public int width = 11;
     public int height = 7;
-    public int offset = 1;
+    public float offset = 1;
 
     public GameObject squarePrefab;
